Made editor fonts more readable. Continued lua experiments. Started to build cge export UI.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@1025 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ tools/anime_studio/cge2.lua
@@ -9,6 +9,9 @@ cam = 0
 visibility = {}
 pictures = {}
 
+dialog = {}
+layer_menu = nil
+
 list_file = nil
 objects_file = nil
 
@@ -89,6 +92,20 @@ function drawLayer(moho, layer)
   vis:SetValue(0, false)
 end
 
+function buildMenu(moho, layer)
+  dialog.layers[count] = MOHO.MSG_BASE + count
+  layer_menu:AddItem(layer:Name(), 0, dialog.layers[count])
+  if (count == 1) then
+    layer_menu:SetChecked(dialog.layers[count], true)
+  end
+  count = count + 1
+end
+
+function handleMessage(dia, message)
+  local check = not layer_menu:IsChecked(message)
+  layer_menu:SetChecked(message, check)
+end
+
 function bla(moho)
   local cge = require("cge")
   count = 0
@@ -97,6 +114,21 @@ function bla(moho)
   view_height = doc:Height()
   cam = doc.fCameraTrack:GetValue(0)
   print("image dimensions:"..view_width.."x"..view_height)
+  
+  dialog.layers = {}
+  local dia = LM.GUI.SimpleDialog("Export Settings", dialog)
+  local popup = LM.GUI.PopupMenu(100, false)  
+  dia:GetLayout():AddChild(popup)
+  layer_menu = LM.GUI.Menu("Layers")
+  popup:SetMenu(layer_menu)
+  
+  count = 1
+  traverseLayers(moho, doc, buildMenu)
+  dia.HandleMessage = handleMessage
+  
+  if dia:DoModal() ~= LM.GUI.MSG_OK then
+    return
+  end
   
   cge.io.removeDir(file_path.."gfx")
   cge.io.createDir(file_path.."gfx")
@@ -107,10 +139,11 @@ function bla(moho)
   objects_file:write(view_height.."\n")
   objects_file:write("0\n0\n5\n10\n3\n\n\n")
   
+  count = 1
   traverseLayers(moho, doc, makeInvisble)
   count = 1
   traverseLayers(moho, doc, drawLayer)
-  count = 0
+  count = 1
   traverseLayers(moho, doc, restoreVisibility)
   
   doc:SetShape(view_width, view_height)
